feat: Add lazy loading for images in ChannelArticleTable
This commit adds lazy loading for images in the ChannelArticleTable component. By using the `react-router-dom` package, the `useParams` hook is imported to retrieve the `id` parameter from the URL. This parameter is then used to generate the URL for the "see more" link in each row. This change improves the performance of the page by loading images only when they are in the viewport, reducing the initial page load time.